# frontend/api/swagger.json
{"openapi":"3.0.0","info":{"title":"TikTok Video Scraper API","version":"1.0.0","description":"API for scraping and downloading TikTok videos","contact":{"name":"API Support"}},"servers":[{"url":"http://localhost:3000","description":"Development server"}],"components":{"schemas":{"Error":{"type":"object","properties":{"success":{"type":"boolean","example":false},"status":{"type":"string","example":"error"},"code":{"type":"integer","example":400},"message":{"type":"string","example":"Invalid request"}}},"VideoUrl":{"type":"object","properties":{"success":{"type":"boolean","example":true},"status":{"type":"string","example":"success"},"data":{"type":"object","properties":{"video_url":{"type":"string","example":"https://v16-webapp.tiktok.com/..."},"video_id":{"type":"string","example":"7153458473287564554"},"metadata":{"type":"object","properties":{"id":{"type":"string","example":"7153458473287564554"},"username":{"type":"string","example":"username"},"description":{"type":"string","example":"Video description"}}}}}}},"VideoDownload":{"type":"object","properties":{"success":{"type":"boolean","example":true},"status":{"type":"string","example":"success"},"data":{"type":"object","properties":{"video_url":{"type":"string","example":"https://v16-webapp.tiktok.com/..."},"file_name":{"type":"string","example":"tiktok_7153458473287564554.mp4"},"download_path":{"type":"string","example":"/downloads/tiktok_7153458473287564554.mp4"},"metadata":{"type":"object","properties":{"id":{"type":"string","example":"7153458473287564554"},"username":{"type":"string","example":"username"},"description":{"type":"string","example":"Video description"}}}}}}}}},"paths":{"/api/tiktok-video":{"get":{"summary":"Returns the direct URL of a TikTok video","description":"Extracts video information and direct URL from a given TikTok video URL","tags":["Video"],"parameters":[{"in":"query","name":"url","required":true,"schema":{"type":"string"},"description":"TikTok video URL"}],"responses":{"200":{"description":"TikTok video URL successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoUrl"}}}},"400":{"description":"Invalid request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/tiktok-download":{"get":{"summary":"Downloads a TikTok video","description":"Downloads a video from the given TikTok video URL and returns its information","tags":["Video"],"parameters":[{"in":"query","name":"url","required":true,"schema":{"type":"string"},"description":"TikTok video URL"}],"responses":{"200":{"description":"TikTok video successfully downloaded","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoDownload"}}}},"400":{"description":"Invalid request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/stream/{filename}":{"get":{"summary":"Streams a downloaded video","description":"Streams a previously downloaded TikTok video","tags":["Stream"],"parameters":[{"in":"path","name":"filename","required":true,"schema":{"type":"string"},"description":"Name of the downloaded video file"}],"responses":{"200":{"description":"Video stream","content":{"video/mp4":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"File not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}}},"tags":[]}